All functions and types

CounterfactualExplanations.Generators.AbstractGenerator

CounterfactualExplanations.Generators.GenericGenerator

CounterfactualExplanations.Generators.GreedyGenerator

CounterfactualExplanations.Models.AbstractFittedModel

CounterfactualExplanations.Models.BayesianLogisticModel

CounterfactualExplanations.Models.LogisticModel

CounterfactualExplanations.Losses.hinge_loss

CounterfactualExplanations.Models.logits

CounterfactualExplanations.Models.logits

CounterfactualExplanations.Models.probs

CounterfactualExplanations.Models.probs

CounterfactualExplanations.generate_counterfactual

Exported functions

CounterfactualExplanations.generate_counterfactual — Method

generate_counterfactual(generator::AbstractGenerator, x::Vector, M::Models.AbstractFittedModel, target::AbstractFloat, γ::AbstractFloat; T=1000)

Takes a recourse generator, the factual sample x, the fitted model M, the target label and its desired threshold probability γ. Returns the generated recourse (an object of type Recourse).

Examples

Generic generator

using CounterfactualExplanations.Models
w = [1.0 -2.0] # true coefficients
b = [0]
x = [-1,0.5]
target = 1.0
γ = 0.9
M = LogisticModel(w, b)
generator = GenericGenerator(0.1,0.1,1e-5,:logitbinarycrossentropy,nothing)
counterfactual = generate_counterfactual(generator, x, M, target, γ); # generate recourse

Greedy generator (Bayesian model only)

using Random, LinearAlgebra
Random.seed!(1234)
μ = [0 1.0 -2.0] # MAP coefficients
Σ = Symmetric(reshape(randn(9),3,3).*0.1 + UniformScaling(1.0)) # MAP covariance matrix
x = [-1,0.5]
target = 1.0
γ = 0.9
M = CounterfactualExplanations.Models.BayesianLogisticModel(μ, Σ);
generator = GreedyGenerator(0.01,20,:logitbinarycrossentropy,nothing)
counterfactual = generate_counterfactual(generator, x, M, target, γ); # generate recourse

See also:

GenericGenerator(λ::AbstractFloat, ϵ::AbstractFloat, τ::AbstractFloat, loss::Symbol, 𝑭::Union{Nothing,Vector{Symbol}})

GreedyGenerator(δ::AbstractFloat, n::Int64, loss::Symbol, 𝑭::Union{Nothing,Vector{Symbol}}).

CounterfactualExplanations.Generators.AbstractGenerator — Type

AbstractGenerator

An abstract type that serves as the base type for recourse generators.

CounterfactualExplanations.Generators.GenericGenerator — Type

GenericGenerator(λ::AbstractFloat, ϵ::AbstractFloat, τ::AbstractFloat, loss::Symbol, 𝑭::Union{Nothing,Vector{Symbol}})

A constructor for a generic recourse generator. It takes values for the complexity penalty λ, the learning rate ϵ, the tolerance for convergence τ, the type of loss function to be used in the recourse objective and a mutability constraint mask 𝑭.

Examples

generator = GenericGenerator(0.1,0.1,1e-5,:logitbinarycrossentropy,nothing)

See also:

generate_counterfactual(generator::AbstractGradientBasedGenerator, x::Vector, M::Models.AbstractFittedModel, target::AbstractFloat; T=1000)

CounterfactualExplanations.Generators.GreedyGenerator — Type

GreedyGenerator(δ::AbstractFloat, n::Int, loss::Symbol, 𝑭::Union{Nothing,Vector{Symbol}})

Constructs a greedy recourse generator for Bayesian models. It takes values for the perturbation size δ, the maximum number of times n that any feature can be changed, the type of loss function to be used in the recourse objective and a mutability constraint mask 𝑭.

Examples

generator = GreedyGenerator(0.01,20,:logitbinarycrossentropy, nothing)

See also:

generate_counterfactual(generator::AbstractGradientBasedGenerator, x::Vector, M::Models.AbstractFittedModel, target::AbstractFloat; T=1000)

CounterfactualExplanations.Models.AbstractFittedModel — Type

AbstractFittedModel

Base type for fitted models.

CounterfactualExplanations.Models.BayesianLogisticModel — Type

BayesianLogisticModel(μ::Matrix,Σ::Matrix)

Constructs a Bayesian logistic classifier based on maximum a posteriori (MAP) estimates μ (coefficients including constant term(s)) and Σ (covariance matrix).

Examples

using Random, LinearAlgebra
Random.seed!(1234)
μ = [0 1.0 -2.0] # MAP coefficients
Σ = Symmetric(reshape(randn(9),3,3).*0.1 + UniformScaling(1.0)) # MAP covariance matrix
M = CounterfactualExplanations.Models.BayesianLogisticModel(μ, Σ);

See also:

logits(M::BayesianLogisticModel, X::AbstractArray)

probs(M::BayesianLogisticModel, X::AbstractArray)

CounterfactualExplanations.Models.LogisticModel — Type

LogisticModel(W::Matrix,b::AbstractArray)

Constructs a logistic classifier based on arrays containing coefficients w and constant terms b.

Examples

w = [1.0 -2.0] # estimated coefficients
b = [0] # estimated constant
M = CounterfactualExplanations.Models.LogisticModel(w, b);

See also:

logits(M::LogisticModel, X::AbstractArray)

probs(M::LogisticModel, X::AbstractArray)

CounterfactualExplanations.Models.logits — Method

logits(M::BayesianLogisticModel, X::AbstractArray)

Computes logits as μ[1ᵀ Xᵀ]ᵀ.

Examples

using CounterfactualExplanations.Models
using Random, LinearAlgebra
Random.seed!(1234)
μ = [0 1.0 -2.0] # MAP coefficients
Σ = Symmetric(reshape(randn(9),3,3).*0.1 + UniformScaling(1.0)) # MAP covariance matrix
M = BayesianLogisticModel(μ, Σ);
x = [1,1]
logits(M, x)

See also BayesianLogisticModel(μ::Matrix,Σ::Matrix)

CounterfactualExplanations.Models.logits — Method

logits(M::LogisticModel, X::AbstractArray)

Computes logits as WX+b.

Examples

using CounterfactualExplanations.Models
w = [1.0 -2.0] # estimated coefficients
b = [0] # estimated constant
M = LogisticModel(w, b);
x = [1,1]
logits(M, x)

See also LogisticModel(W::Matrix,b::AbstractArray).

CounterfactualExplanations.Models.probs — Method

probs(M::BayesianLogisticModel, X::AbstractArray)

Computes predictive probabilities using a Probit approximation.

Examples

using CounterfactualExplanations.Models
using Random, LinearAlgebra
Random.seed!(1234)
μ = [0 1.0 -2.0] # MAP coefficients
Σ = Symmetric(reshape(randn(9),3,3).*0.1 + UniformScaling(1.0)) # MAP covariance matrix
M = BayesianLogisticModel(μ, Σ);
x = [1,1]
probs(M, x)

See also BayesianLogisticModel(μ::Matrix,Σ::Matrix)

CounterfactualExplanations.Models.probs — Method

probs(M::LogisticModel, X::AbstractArray)

Computes predictive probabilities from logits as σ(WX+b) where 'σ' is the sigmoid function.

Examples

using CounterfactualExplanations.Models
w = [1.0 -2.0] # estimated coefficients
b = [0] # estimated constant
M = LogisticModel(w, b);
x = [1,1]
probs(M, x)

See also LogisticModel(W::Matrix,b::AbstractArray).

CounterfactualExplanations.Losses.hinge_loss — Method

hinge_loss(ŷ, y)

Uses Flux.Losses.hinge_loss to compute hinge loss. If provided labels are in {0,1} they are mapped to {-1,1}.

Internal functions