All functions and types

CounterfactualExplanations.DataPreprocessing.CounterfactualData

CounterfactualExplanations.Generators.AbstractGenerator

CounterfactualExplanations.Generators.AbstractGradientBasedGenerator

CounterfactualExplanations.Generators.GenericGenerator

CounterfactualExplanations.Generators.GreedyGenerator

CounterfactualExplanations.Models.AbstractFittedModel

CounterfactualExplanations.Models.BayesianLogisticModel

CounterfactualExplanations.Models.LogisticModel

CounterfactualExplanations.Counterfactuals.apply_mutability

CounterfactualExplanations.Counterfactuals.converged

CounterfactualExplanations.Counterfactuals.counterfactual

CounterfactualExplanations.Counterfactuals.counterfactual_label

CounterfactualExplanations.Counterfactuals.counterfactual_probability

CounterfactualExplanations.Counterfactuals.encode_target

CounterfactualExplanations.Counterfactuals.factual

CounterfactualExplanations.Counterfactuals.factual_label

CounterfactualExplanations.Counterfactuals.factual_probability

CounterfactualExplanations.Counterfactuals.get_counterfactual_state

CounterfactualExplanations.Counterfactuals.initialize!

CounterfactualExplanations.Counterfactuals.output_dim

CounterfactualExplanations.Counterfactuals.path

CounterfactualExplanations.Counterfactuals.steps_exhausted

CounterfactualExplanations.Counterfactuals.target_probs

CounterfactualExplanations.Counterfactuals.terminated

CounterfactualExplanations.Counterfactuals.threshold_reached

CounterfactualExplanations.Counterfactuals.total_steps

CounterfactualExplanations.Counterfactuals.update!

CounterfactualExplanations.Data.cats_dogs_laplace

CounterfactualExplanations.Data.cats_dogs_model

CounterfactualExplanations.Data.mnist_data

CounterfactualExplanations.Data.mnist_ensemble

CounterfactualExplanations.Data.mnist_model

CounterfactualExplanations.Data.toy_data_linear

CounterfactualExplanations.Data.toy_data_multi

CounterfactualExplanations.Data.toy_data_non_linear

CounterfactualExplanations.Data.ucr_data

CounterfactualExplanations.Data.ucr_model

CounterfactualExplanations.DataPreprocessing.apply_domain_constraints

CounterfactualExplanations.DataPreprocessing.mutability_constraints

CounterfactualExplanations.DataPreprocessing.select_factual

CounterfactualExplanations.Generators.conditions_satisified

CounterfactualExplanations.Generators.conditions_satisified

CounterfactualExplanations.Generators.generate_perturbations

CounterfactualExplanations.Generators.generate_perturbations

CounterfactualExplanations.Generators.h

CounterfactualExplanations.Generators.mutability_constraints

CounterfactualExplanations.Generators.mutability_constraints

CounterfactualExplanations.Generators.ℓ

CounterfactualExplanations.Generators.∂h

CounterfactualExplanations.Generators.∂ℓ

CounterfactualExplanations.Generators.∇

CounterfactualExplanations.Generators.∇

CounterfactualExplanations.Models.logits

CounterfactualExplanations.Models.logits

CounterfactualExplanations.Models.probs

CounterfactualExplanations.Models.probs

CounterfactualExplanations.generate_counterfactual

Flux.Losses.add_blanks

Flux.Losses.binary_focal_loss

Flux.Losses.binarycrossentropy

Flux.Losses.crossentropy

Flux.Losses.ctc_loss

Flux.Losses.dice_coeff_loss

Flux.Losses.focal_loss

Flux.Losses.hinge_loss

Flux.Losses.huber_loss

Flux.Losses.kldivergence

Flux.Losses.label_smoothing

Flux.Losses.logaddexp

Flux.Losses.logitbinarycrossentropy

Flux.Losses.logitcrossentropy

Flux.Losses.mae

Flux.Losses.mse

Flux.Losses.msle

Flux.Losses.poisson_loss

Flux.Losses.squared_hinge_loss

Flux.Losses.tversky_loss

Flux.Losses.xlogx

Flux.Losses.xlogy

Exported functions

CounterfactualExplanations.generate_counterfactual — Method

generate_counterfactual(
    x::Union{AbstractArray,Int}, target::Union{AbstractFloat,Int}, data::CounterfactualData, M::Models.AbstractFittedModel, generator::AbstractGenerator;
    γ::AbstractFloat=0.75, T=1000
)

The core function that is used to run counterfactual search for a given factual x, target, counterfactual data, model and generator. Keywords can be used to specify the desired threshold for the predicted target class probability and the maximum number of iterations.

Examples

Generic generator

using CounterfactualExplanations

# Data:
using CounterfactualExplanations.Data
using Random
Random.seed!(1234)
xs, ys = Data.toy_data_linear()
X = hcat(xs...)
counterfactual_data = CounterfactualData(X,ys')

# Model
using CounterfactualExplanations.Models: LogisticModel, probs 
# Logit model:
w = [1.0 1.0] # true coefficients
b = 0
M = LogisticModel(w, [b])

# Randomly selected factual:
x = select_factual(counterfactual_data,rand(1:size(X)[2]))
y = round(probs(M, x)[1])
target = round(probs(M, x)[1])==0 ? 1 : 0 

# Counterfactual search:
generator = GenericGenerator()
counterfactual = generate_counterfactual(x, target, counterfactual_data, M, generator)

CounterfactualExplanations.DataPreprocessing.CounterfactualData — Method

CounterfactualData(
    X::AbstractArray, y::AbstractArray;
    mutability::Union{Vector{Symbol},Nothing}=nothing,
    domain::Union{Any,Nothing}=nothing,
    categorical::Union{Vector{Int},Nothing}=nothing,
    continuous::Union{Vector{Int},Nothing}=nothing,
)

This outer constructor method prepares features X and labels y to be used with the package. Mutability and domain constraints can be added for the features. The function also accepts arguments that specify which features are categorical and which are continues. These arguments are currently not used.

Examples

using CounterfactualExplanations.Data
x, y = toy_data_linear()
X = hcat(x...)
counterfactual_data = CounterfactualData(X,y')

CounterfactualExplanations.DataPreprocessing.apply_domain_constraints — Method

apply_domain_constraints(counterfactual_data::CounterfactualData, x::AbstractArray)

A subroutine that is used to apply the predetermined domain constraints.

CounterfactualExplanations.DataPreprocessing.select_factual — Method

select_factual(counterfactual_data::CounterfactualData, index::Int)

A convenience method that can be used to access the the feature matrix.

CounterfactualExplanations.Counterfactuals.converged — Method

converged(counterfactual_explanation::CounterfactualExplanation)

A convenience method to determine if the counterfactual search has converged.

CounterfactualExplanations.Counterfactuals.initialize! — Method

initialize!(counterfactual_explanation::CounterfactualExplanation)

A subroutine that intializes the counterfactual search.

CounterfactualExplanations.Counterfactuals.path — Method

path(counterfactual_explanation::CounterfactualExplanation)

A convenience method that returns the entire counterfactual path.

CounterfactualExplanations.Counterfactuals.target_probs — Function

target_probs(counterfactual_explanation::CounterfactualExplanation, x::Union{AbstractArray, Nothing}=nothing)

Returns the predicted probability of the target class for x. If x is nothing, the predicted probability corresponding to the counterfactual value is returned.

CounterfactualExplanations.Counterfactuals.terminated — Method

terminated(counterfactual_explanation::CounterfactualExplanation)

A convenience method to determine if the counterfactual search has terminated.

CounterfactualExplanations.Counterfactuals.total_steps — Method

total_steps(counterfactual_explanation::CounterfactualExplanation)

A convenience method that returns the total number of steps of the counterfactual search.

CounterfactualExplanations.Counterfactuals.update! — Method

update!(counterfactual_explanation::CounterfactualExplanation)

An important subroutine that updates the counterfactual explanation. It takes a snapshot of the current counterfactual search state and passes it to the generator. Based on the current state the generator generates perturbations. Various constraints are then applied to the proposed vector of feature perturbations. Finally, the counterfactual search state is updated.

CounterfactualExplanations.Models.AbstractFittedModel — Type

AbstractFittedModel

Base type for fitted models.

CounterfactualExplanations.Models.BayesianLogisticModel — Type

BayesianLogisticModel(μ::Matrix,Σ::Matrix)

Constructs a Bayesian logistic classifier based on maximum a posteriori (MAP) estimates μ (coefficients including constant term(s)) and Σ (covariance matrix).

Examples

using Random, LinearAlgebra
Random.seed!(1234)
μ = [0 1.0 -2.0] # MAP coefficients
Σ = Symmetric(reshape(randn(9),3,3).*0.1 + UniformScaling(1.0)) # MAP covariance matrix
M = CounterfactualExplanations.Models.BayesianLogisticModel(μ, Σ);

See also:

logits(M::BayesianLogisticModel, X::AbstractArray)

probs(M::BayesianLogisticModel, X::AbstractArray)

CounterfactualExplanations.Models.LogisticModel — Type

LogisticModel(W::Matrix,b::AbstractArray)

Constructs a logistic classifier based on arrays containing coefficients w and constant terms b.

Examples

w = [1.0 -2.0] # estimated coefficients
b = [0] # estimated constant
M = CounterfactualExplanations.Models.LogisticModel(w, b);

See also:

logits(M::LogisticModel, X::AbstractArray)

probs(M::LogisticModel, X::AbstractArray)

CounterfactualExplanations.Models.logits — Method

logits(M::BayesianLogisticModel, X::AbstractArray)

Computes logits as μ[1ᵀ Xᵀ]ᵀ.

Examples

using CounterfactualExplanations.Models
using Random, LinearAlgebra
Random.seed!(1234)
μ = [0 1.0 -2.0] # MAP coefficients
Σ = Symmetric(reshape(randn(9),3,3).*0.1 + UniformScaling(1.0)) # MAP covariance matrix
M = BayesianLogisticModel(μ, Σ);
x = [1,1]
logits(M, x)

See also BayesianLogisticModel(μ::Matrix,Σ::Matrix)

CounterfactualExplanations.Models.logits — Method

logits(M::LogisticModel, X::AbstractArray)

Computes logits as WX+b.

Examples

using CounterfactualExplanations.Models
w = [1.0 -2.0] # estimated coefficients
b = [0] # estimated constant
M = LogisticModel(w, b);
x = [1,1]
logits(M, x)

See also LogisticModel(W::Matrix,b::AbstractArray).

CounterfactualExplanations.Models.probs — Method

probs(M::BayesianLogisticModel, X::AbstractArray)

Computes predictive probabilities using a Probit approximation.

Examples

using CounterfactualExplanations.Models
using Random, LinearAlgebra
Random.seed!(1234)
μ = [0 1.0 -2.0] # MAP coefficients
Σ = Symmetric(reshape(randn(9),3,3).*0.1 + UniformScaling(1.0)) # MAP covariance matrix
M = BayesianLogisticModel(μ, Σ);
x = [1,1]
probs(M, x)

See also BayesianLogisticModel(μ::Matrix,Σ::Matrix)

CounterfactualExplanations.Models.probs — Method

probs(M::LogisticModel, X::AbstractArray)

Computes predictive probabilities from logits as σ(WX+b) where 'σ' is the sigmoid function.

Examples

using CounterfactualExplanations.Models
w = [1.0 -2.0] # estimated coefficients
b = [0] # estimated constant
M = LogisticModel(w, b);
x = [1,1]
probs(M, x)

See also LogisticModel(W::Matrix,b::AbstractArray).

CounterfactualExplanations.Generators.AbstractGenerator — Type

AbstractGenerator

An abstract type that serves as the base type for counterfactual generators.

CounterfactualExplanations.Generators.AbstractGradientBasedGenerator — Type

AbstractGradientBasedGenerator

An abstract type that serves as the base type for gradient-based counterfactual generators.

CounterfactualExplanations.Generators.GenericGenerator — Method

GenericGenerator(
    ;
    loss::Symbol=:logitbinarycrossentropy,
    complexity::Function=norm,
    mutability::Union{Nothing,Vector{Symbol}}=nothing,
    λ::AbstractFloat=0.1,
    ϵ::AbstractFloat=0.1,
    τ::AbstractFloat=1e-5
)

An outer constructor method that instantiates a generic generator.

Examples

generator = GenericGenerator()

CounterfactualExplanations.Generators.GreedyGenerator — Method

GreedyGenerator(
    ;
    loss::Symbol=:logitbinarycrossentropy,
    mutability::Union{Nothing,Vector{Symbol}}=nothing,
    δ::Union{AbstractFloat,Nothing}=nothing,
    n::Union{Int,Nothing}=nothing
)

An outer constructor method that instantiates a greedy generator.

Examples

generator = GreedyGenerator()

CounterfactualExplanations.Generators.conditions_satisified — Method

conditions_satisified(generator::AbstractGradientBasedGenerator, counterfactual_state::CounterfactualState)

The default method to check if the all conditions for convergence of the counterfactual search have been satisified for gradient-based generators.

CounterfactualExplanations.Generators.conditions_satisified — Method

conditions_satisified(generator::GreedyGenerator, counterfactual_state::CounterfactualState)

If all features have been perturbed n times already, then the search terminates.

CounterfactualExplanations.Generators.generate_perturbations — Method

generate_perturbations(generator::AbstractGradientBasedGenerator, counterfactual_state::CounterfactualState)

The default method to generate feature perturbations for gradient-based generators through simple gradient descent.

CounterfactualExplanations.Generators.generate_perturbations — Method

generate_perturbations(generator::GreedyGenerator, counterfactual_state::CounterfactualState)

The default method to generate perturbations for a greedy generator. Only the most salient feature is perturbed.

CounterfactualExplanations.Generators.mutability_constraints — Method

mutability_constraints(generator::AbstractGradientBasedGenerator, counterfactual_state::CounterfactualState)

The default method to return mutability constraints that are dependent on the current counterfactual search state. For generic gradient-based generators, no state-dependent constraints are added.

CounterfactualExplanations.Generators.mutability_constraints — Method

mutability_constraints(generator::GreedyGenerator, counterfactual_state::CounterfactualState)

The default method to return search state dependent mutability constraints for a greedy generator. Features that have been perturbed n times already can no longer be perturbed.

Flux.Losses.binary_focal_loss — Method

binary_focal_loss(ŷ, y; agg=mean, γ=2, ϵ=eps(ŷ))

Return the binaryfocalloss The input, 'ŷ', is expected to be normalized (i.e. softmax output).

For γ == 0, the loss is mathematically equivalent to Losses.binarycrossentropy.

Example

julia> y = [0  1  0
            1  0  1]
2×3 Matrix{Int64}:
 0  1  0
 1  0  1

julia> ŷ = [0.268941  0.5  0.268941
            0.731059  0.5  0.731059]
2×3 Matrix{Float64}:
 0.268941  0.5  0.268941
 0.731059  0.5  0.731059

julia> Flux.binary_focal_loss(ŷ, y) ≈ 0.0728675615927385
true

See also: Losses.focal_loss for multi-class setting

Flux.Losses.binarycrossentropy — Method

binarycrossentropy(ŷ, y; agg = mean, ϵ = eps(ŷ))

Return the binary cross-entropy loss, computed as

agg(@.(-y * log(ŷ + ϵ) - (1 - y) * log(1 - ŷ + ϵ)))

Where typically, the prediction ŷ is given by the output of a sigmoid activation. The ϵ term is included to avoid infinity. Using logitbinarycrossentropy is recomended over binarycrossentropy for numerical stability.

Use label_smoothing to smooth the y value as preprocessing before computing the loss.

See also: crossentropy, logitcrossentropy.

Examples

julia> y_bin = Bool[1,0,1]
3-element Vector{Bool}:
 1
 0
 1

julia> y_prob = softmax(reshape(vcat(1:3, 3:5), 2, 3) .* 1f0)
2×3 Matrix{Float32}:
 0.268941  0.5  0.268941
 0.731059  0.5  0.731059

julia> Flux.binarycrossentropy(y_prob[2,:], y_bin)
0.43989f0

julia> all(p -> 0 < p < 1, y_prob[2,:])  # else DomainError
true

julia> y_hot = Flux.onehotbatch(y_bin, 0:1)
2×3 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
 ⋅  1  ⋅
 1  ⋅  1

julia> Flux.crossentropy(y_prob, y_hot)
0.43989f0

Flux.Losses.crossentropy — Method

crossentropy(ŷ, y; dims = 1, ϵ = eps(ŷ), agg = mean)

Return the cross entropy between the given probability distributions; calculated as

agg(-sum(y .* log.(ŷ .+ ϵ); dims))

Cross entropy is typically used as a loss in multi-class classification, in which case the labels y are given in a one-hot format. dims specifies the dimension (or the dimensions) containing the class probabilities. The prediction ŷ is supposed to sum to one across dims, as would be the case with the output of a softmax operation.

For numerical stability, it is recommended to use logitcrossentropy rather than softmax followed by crossentropy .

Use label_smoothing to smooth the true labels as preprocessing before computing the loss.

See also: logitcrossentropy, binarycrossentropy, logitbinarycrossentropy.

Example

julia> y_label = Flux.onehotbatch([0, 1, 2, 1, 0], 0:2)
3×5 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
 1  ⋅  ⋅  ⋅  1
 ⋅  1  ⋅  1  ⋅
 ⋅  ⋅  1  ⋅  ⋅

julia> y_model = softmax(reshape(-7:7, 3, 5) .* 1f0)
3×5 Matrix{Float32}:
 0.0900306  0.0900306  0.0900306  0.0900306  0.0900306
 0.244728   0.244728   0.244728   0.244728   0.244728
 0.665241   0.665241   0.665241   0.665241   0.665241

julia> sum(y_model; dims=1)
1×5 Matrix{Float32}:
 1.0  1.0  1.0  1.0  1.0

julia> Flux.crossentropy(y_model, y_label)
1.6076053f0

julia> 5 * ans ≈ Flux.crossentropy(y_model, y_label; agg=sum)
true

julia> y_smooth = Flux.label_smoothing(y_label, 0.15f0)
3×5 Matrix{Float32}:
 0.9   0.05  0.05  0.05  0.9
 0.05  0.9   0.05  0.9   0.05
 0.05  0.05  0.9   0.05  0.05

julia> Flux.crossentropy(y_model, y_smooth)
1.5776052f0

Flux.Losses.ctc_loss — Method

ctc_loss(ŷ, y)

Computes the connectionist temporal classification loss between ŷ and y.

ŷ must be a classes-by-time matrices, i.e., each row represents a class and each column represents a time step. Additionally, the logsoftmax function will be applied to ŷ, so ŷ must be the raw activation values from the neural network and not, for example, the activations after being passed through a softmax activation function. y must be a 1D array of the labels associated with ŷ. The blank label is assumed to be the last label category in ŷ, so it is equivalent to size(ŷ, 1).

Used for sequence-to-sequence classification problems such as speech recognition and handwriting recognition where the exact time-alignment of the output (e.g., letters) is not needed to solve the problem. See Graves et al. (2006) or Graves (2012) for mathematical details.

Flux.Losses.dice_coeff_loss — Method

dice_coeff_loss(ŷ, y; smooth = 1)

Return a loss based on the dice coefficient. Used in the V-Net image segmentation architecture. Similar to the F1_score. Calculated as:

1 - 2*sum(|ŷ .* y| + smooth) / (sum(ŷ.^2) + sum(y.^2) + smooth)

Flux.Losses.focal_loss — Method

focal_loss(ŷ, y; dims=1, agg=mean, γ=2, ϵ=eps(ŷ))

Return the focal_loss which can be used in classification tasks with highly imbalanced classes. It down-weights well-classified examples and focuses on hard examples. The input, 'ŷ', is expected to be normalized (i.e. softmax output).

The modulating factor, γ, controls the down-weighting strength. For γ == 0, the loss is mathematically equivalent to Losses.crossentropy.

Example

julia> y = [1  0  0  0  1
            0  1  0  1  0
            0  0  1  0  0]
3×5 Matrix{Int64}:
 1  0  0  0  1
 0  1  0  1  0
 0  0  1  0  0

julia> ŷ = softmax(reshape(-7:7, 3, 5) .* 1f0)
3×5 Matrix{Float32}:
 0.0900306  0.0900306  0.0900306  0.0900306  0.0900306
 0.244728   0.244728   0.244728   0.244728   0.244728
 0.665241   0.665241   0.665241   0.665241   0.665241

julia> Flux.focal_loss(ŷ, y) ≈ 1.1277571935622628
true

See also: Losses.binary_focal_loss for binary (not one-hot) labels

Flux.Losses.hinge_loss — Method

hinge_loss(ŷ, y; agg = mean)

Return the hinge_loss loss given the prediction ŷ and true labels y (containing 1 or -1); calculated as sum(max.(0, 1 .- ŷ .* y)) / size(y, 2).

See also: squared_hinge_loss

Flux.Losses.huber_loss — Method

huber_loss(ŷ, y; δ = 1, agg = mean)

Return the mean of the Huber loss given the prediction ŷ and true values y.

             | 0.5 * |ŷ - y|^2,            for |ŷ - y| <= δ
Huber loss = |
             |  δ * (|ŷ - y| - 0.5 * δ), otherwise

Flux.Losses.kldivergence — Method

kldivergence(ŷ, y; agg = mean, ϵ = eps(ŷ))

Return the Kullback-Leibler divergence between the given probability distributions.

The KL divergence is a measure of how much one probability distribution is different from the other. It is always non-negative, and zero only when both the distributions are equal.

Example

julia> p1 = [1 0; 0 1]
2×2 Matrix{Int64}:
 1  0
 0  1

julia> p2 = fill(0.5, 2, 2)
2×2 Matrix{Float64}:
 0.5  0.5
 0.5  0.5

julia> Flux.kldivergence(p2, p1) ≈ log(2)
true

julia> Flux.kldivergence(p2, p1; agg = sum) ≈ 2log(2)
true

julia> Flux.kldivergence(p2, p2; ϵ = 0)  # about -2e-16 with the regulator
0.0

julia> Flux.kldivergence(p1, p2; ϵ = 0)  # about 17.3 with the regulator
Inf

Flux.Losses.label_smoothing — Method

label_smoothing(y::Union{Number, AbstractArray}, α; dims::Int=1)

Returns smoothed labels, meaning the confidence on label values are relaxed.

When y is given as one-hot vector or batch of one-hot, its calculated as

y .* (1 - α) .+ α / size(y, dims)

when y is given as a number or batch of numbers for binary classification, its calculated as

y .* (1 - α) .+ α / 2

in which case the labels are squeezed towards 0.5.

α is a number in interval (0, 1) called the smoothing factor. Higher the value of α larger the smoothing of y.

dims denotes the one-hot dimension, unless dims=0 which denotes the application of label smoothing to binary distributions encoded in a single number.

Example

julia> y = Flux.onehotbatch([1, 1, 1, 0, 1, 0], 0:1)
2×6 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
 ⋅  ⋅  ⋅  1  ⋅  1
 1  1  1  ⋅  1  ⋅

julia> y_smoothed = Flux.label_smoothing(y, 0.2f0)
2×6 Matrix{Float32}:
 0.1  0.1  0.1  0.9  0.1  0.9
 0.9  0.9  0.9  0.1  0.9  0.1

julia> y_sim = softmax(y .* log(2f0))
2×6 Matrix{Float32}:
 0.333333  0.333333  0.333333  0.666667  0.333333  0.666667
 0.666667  0.666667  0.666667  0.333333  0.666667  0.333333

julia> y_dis = vcat(y_sim[2,:]', y_sim[1,:]')
2×6 Matrix{Float32}:
 0.666667  0.666667  0.666667  0.333333  0.666667  0.333333
 0.333333  0.333333  0.333333  0.666667  0.333333  0.666667

julia> Flux.crossentropy(y_sim, y) < Flux.crossentropy(y_sim, y_smoothed)
true

julia> Flux.crossentropy(y_dis, y) > Flux.crossentropy(y_dis, y_smoothed)
true

Flux.Losses.logitbinarycrossentropy — Method

logitbinarycrossentropy(ŷ, y; agg = mean)

Mathematically equivalent to binarycrossentropy(σ(ŷ), y) but is more numerically stable.

See also: crossentropy, logitcrossentropy.

Examples

julia> y_bin = Bool[1,0,1];

julia> y_model = Float32[2, -1, pi]
3-element Vector{Float32}:
  2.0
 -1.0
  3.1415927

julia> Flux.logitbinarycrossentropy(y_model, y_bin)
0.160832f0

julia> Flux.binarycrossentropy(sigmoid.(y_model), y_bin)
0.16083185f0

Flux.Losses.logitcrossentropy — Method

logitcrossentropy(ŷ, y; dims = 1, agg = mean)

Return the cross entropy calculated by

agg(-sum(y .* logsoftmax(ŷ; dims); dims))

This is mathematically equivalent to crossentropy(softmax(ŷ), y), but is more numerically stable than using functions crossentropy and softmax separately.

See also: binarycrossentropy, logitbinarycrossentropy, label_smoothing.

Example

julia> y_label = Flux.onehotbatch(collect("abcabaa"), 'a':'c')
3×7 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
 1  ⋅  ⋅  1  ⋅  1  1
 ⋅  1  ⋅  ⋅  1  ⋅  ⋅
 ⋅  ⋅  1  ⋅  ⋅  ⋅  ⋅

julia> y_model = reshape(vcat(-9:0, 0:9, 7.5f0), 3, 7)
3×7 Matrix{Float32}:
 -9.0  -6.0  -3.0  0.0  2.0  5.0  8.0
 -8.0  -5.0  -2.0  0.0  3.0  6.0  9.0
 -7.0  -4.0  -1.0  1.0  4.0  7.0  7.5

julia> Flux.logitcrossentropy(y_model, y_label)
1.5791205f0

julia> Flux.crossentropy(softmax(y_model), y_label)
1.5791197f0

Flux.Losses.mae — Method

mae(ŷ, y; agg = mean)

Return the loss corresponding to mean absolute error:

agg(abs.(ŷ .- y))

Example

julia> y_model = [1.1, 1.9, 3.1];

julia> Flux.mae(y_model, 1:3)
0.10000000000000009

Flux.Losses.mse — Method

mse(ŷ, y; agg = mean)

Return the loss corresponding to mean square error:

agg((ŷ .- y) .^ 2)

See also: mae, msle, crossentropy.

Example

julia> y_model = [1.1, 1.9, 3.1];

julia> y_true = 1:3;

julia> Flux.mse(y_model, y_true)
0.010000000000000018

Flux.Losses.msle — Method

msle(ŷ, y; agg = mean, ϵ = eps(ŷ))

The loss corresponding to mean squared logarithmic errors, calculated as

agg((log.(ŷ .+ ϵ) .- log.(y .+ ϵ)) .^ 2)

The ϵ term provides numerical stability. Penalizes an under-estimation more than an over-estimatation.

Example

julia> Flux.msle(Float32[1.1, 2.2, 3.3], 1:3)
0.009084041f0

julia> Flux.msle(Float32[0.9, 1.8, 2.7], 1:3)
0.011100831f0

Flux.Losses.poisson_loss — Method

poisson_loss(ŷ, y)

Return how much the predicted distribution ŷ diverges from the expected Poisson

distribution y; calculated as sum(ŷ .- y .* log.(ŷ)) / size(y, 2).

More information..

Flux.Losses.squared_hinge_loss — Method

squared_hinge_loss(ŷ, y)

Return the squared hinge_loss loss given the prediction ŷ and true labels y (containing 1 or -1); calculated as sum((max.(0, 1 .- ŷ .* y)).^2) / size(y, 2).

See also: hinge_loss

Flux.Losses.tversky_loss — Method

tversky_loss(ŷ, y; β = 0.7)

Return the Tversky loss. Used with imbalanced data to give more weight to false negatives. Larger β weigh recall more than precision (by placing more emphasis on false negatives) Calculated as: 1 - sum(|y .* ŷ| + 1) / (sum(y .* ŷ + β(1 .- y) . ŷ + (1 - β)y . (1 .- ŷ)) + 1)

Internal functions

CounterfactualExplanations.Data.cats_dogs_laplace — Method

cats_dogs_laplace()

A convenience function to load the pre-trained MLP with Laplace approximation to classify cats and dogs.

CounterfactualExplanations.Data.cats_dogs_model — Method

cats_dogs_model()

A convenience function to load the pre-trained MLP to classify cats and dogs.

CounterfactualExplanations.Data.mnist_data — Method

mnist_data()

A convenience function to load MNIST training data.

CounterfactualExplanations.Data.mnist_ensemble — Method

mnist_ensemble()

A convenience function to load the pre-trained deep ensemble of MLPs for MNIST training data.

CounterfactualExplanations.Data.mnist_model — Method

mnist_model()

A convenience function to load the pre-trained MLP for MNIST training data.

CounterfactualExplanations.Data.toy_data_linear — Function

toy_data_linear(N=100)

A convenience function to load linearly separable synthetic data.

Examples

toy_data_linear()

CounterfactualExplanations.Data.toy_data_multi — Function

toy_data_multi(N=100)

A convenience function to load multi-class synthetic data.

Examples

toy_data_multi()

CounterfactualExplanations.Data.toy_data_non_linear — Function

toy_data_non_linear(N=100)

A convenience function to load synthetic data that are not linearly separable.

Examples

toy_data_non_linear()

CounterfactualExplanations.Data.ucr_data — Method

ucr_data()

A convenience function to load UCR time series data.

CounterfactualExplanations.Data.ucr_model — Method

ucr_model()

A convenience function to load a pre-trained classification model for the UCR time series data.

CounterfactualExplanations.DataPreprocessing.mutability_constraints — Method

mutability_constraints(counterfactual_data::CounterfactualData)

A convience function that returns the mutability constraints. If none were specified, it is assumed that all features are mutable in :both directions.

CounterfactualExplanations.Counterfactuals.apply_mutability — Method

apply_mutability(Δx′::AbstractArray, counterfactual_explanation::CounterfactualExplanation)

A subroutine that applies mutability constraints to the proposed vector of feature perturbations.

CounterfactualExplanations.Counterfactuals.counterfactual — Method

counterfactual(counterfactual_explanation::CounterfactualExplanation)

A convenience method to get the counterfactual value.

CounterfactualExplanations.Counterfactuals.counterfactual_label — Method

counterfactual_label(counterfactual_explanation::CounterfactualExplanation)

A convenience method to get the predicted label associated with the counterfactual value.

CounterfactualExplanations.Counterfactuals.counterfactual_probability — Method

counterfactual_probability(counterfactual_explanation::CounterfactualExplanation)

A convenience method to compute the class probabilities of the counterfactual value.

CounterfactualExplanations.Counterfactuals.encode_target — Method

encode_target(counterfactual_explanation::CounterfactualExplanation)

A convenience method to encode the target variable, if necessary.

CounterfactualExplanations.Counterfactuals.factual — Method

factual(counterfactual_explanation::CounterfactualExplanation)

A convenience method to get the factual value.

CounterfactualExplanations.Counterfactuals.factual_label — Method

factual_label(counterfactual_explanation::CounterfactualExplanation)

A convenience method to get the predicted label associated with the factual value.

CounterfactualExplanations.Counterfactuals.factual_probability — Method

factual_probability(counterfactual_explanation::CounterfactualExplanation)

A convenience method to compute the class probabilities of the factual value.

CounterfactualExplanations.Counterfactuals.get_counterfactual_state — Method

get_counterfactual_state(counterfactual_explanation::CounterfactualExplanation)

A subroutine that is used to take a snapshot of the current counterfactual search state. This snapshot is passed to the counterfactual generator.

CounterfactualExplanations.Counterfactuals.output_dim — Method

output_dim(counterfactual_explanation::CounterfactualExplanation)

A convenience method that computes the output dimension of the predictive model.

CounterfactualExplanations.Counterfactuals.steps_exhausted — Method

steps_exhausted(counterfactual_explanation::CounterfactualExplanation)

A convenience method that checks if the number of maximum iterations has been exhausted.

CounterfactualExplanations.Counterfactuals.threshold_reached — Method

threshold_reached(counterfactual_explanation::CounterfactualExplanation)

A convenience method that determines of the predefined threshold for the target class probability has been reached.

CounterfactualExplanations.Generators.h — Method

h(generator::AbstractGenerator, counterfactual_state::CounterfactualState)

The default method to apply the generator complexity penalty to the current counterfactual state for any generator.

CounterfactualExplanations.Generators.ℓ — Method

ℓ(generator::AbstractGenerator, counterfactual_state::CounterfactualState)

The default method to apply the generator loss function to the current counterfactual state for any generator.

CounterfactualExplanations.Generators.∂h — Method

∂h(generator::AbstractGradientBasedGenerator, counterfactual_state::CounterfactualState)

The default method to compute the gradient of the complexity penalty at the current counterfactual state for gradient-based generators. It assumes that Zygote.jl has gradient access.

CounterfactualExplanations.Generators.∂ℓ — Method

∂ℓ(generator::AbstractGradientBasedGenerator, counterfactual_state::CounterfactualState)

The default method to compute the gradient of the loss function at the current counterfactual state for gradient-based generators. It assumes that Zygote.jl has gradient access.

CounterfactualExplanations.Generators.∇ — Method

∇(generator::AbstractGradientBasedGenerator, counterfactual_state::CounterfactualState)

The default method to compute the gradient of the counterfactual search objective for gradient-based generators. It simply computes the weighted sum over partial derivates. It assumes that Zygote.jl has gradient access.

CounterfactualExplanations.Generators.∇ — Method

∇(generator::GreedyGenerator, counterfactual_state::CounterfactualState)

he default method to compute the gradient of the counterfactual search objective for a greedy generator. Since no complexity penalty is needed, this gradients just correponds to the partial derivative with respect to the loss function.

Flux.Losses.add_blanks — Method

add_blanks(z)

Adds blanks to the start and end of z, and between items in z

Flux.Losses.logaddexp — Method

logaddexp(a, b) Adds log-space a and b such that the result equals log(exp(a)+exp(b))

Flux.Losses.xlogx — Method

xlogx(x)

Return x * log(x) for x ≥ 0, handling x == 0 by taking the limit from above, to get zero.

Flux.Losses.xlogy — Method

xlogy(x, y)

Return x * log(y) for y > 0, and zero when x == 0.