Reference

In this reference, you will find a detailed overview of the package API.

Reference guides are technical descriptions of the machinery and how to operate it. Reference material is information-oriented.

— Diátaxis

In other words, you come here because you want to take a very close look at the code 🧐.

Content

Reference

Content

Exported functions

Internal functions

Exported functions

CounterfactualExplanations.RawOutputArrayType — Type

RawOutputArrayType

A type union for the allowed type for the output array y.

CounterfactualExplanations.RawTargetType — Type

RawTargetType

A type union for the allowed types for the target variable.

CounterfactualExplanations.flux_training_params — Constant

flux_training_params

The default training parameter for FluxModels etc.

CounterfactualExplanations.CounterfactualExplanation — Type

A struct that collects all information relevant to a specific counterfactual explanation for a single individual.

CounterfactualExplanations.CounterfactualExplanation — Method

function CounterfactualExplanation(;
    x::AbstractArray,
    target::RawTargetType,
    data::CounterfactualData,
    M::Models.AbstractFittedModel,
    generator::Generators.AbstractGenerator,
    T::Int = 100,
    latent_space::Union{Nothing,Bool} = nothing,
    num_counterfactuals::Int = 1,
    initialization::Symbol = :add_perturbation,
    generative_model_params::NamedTuple = (;),
)

Outer method to construct a CounterfactualExplanation structure.

CounterfactualExplanations.EncodedOutputArrayType — Type

EncodedOutputArrayType

Type of encoded output array.

CounterfactualExplanations.EncodedTargetType — Type

EncodedTargetType

Type of encoded target variable.

CounterfactualExplanations.OutputEncoder — Type

OutputEncoder

The OutputEncoder takes a raw output array (y) and encodes it.

CounterfactualExplanations.OutputEncoder — Method

(encoder::OutputEncoder)(ynew::RawTargetType)

When called on a new value ynew, the OutputEncoder encodes it based on the initial encoding.

CounterfactualExplanations.OutputEncoder — Method

(encoder::OutputEncoder)()

On call, the OutputEncoder returns the encoded output array.

CounterfactualExplanations.animate_path — Function

animate_path(counterfactual_explanation::CounterfactualExplanation, path=tempdir(); plot_proba::Bool=false, kwargs...)

Returns and animation of the counterfactual path.

Examples

# Search:
generator = GenericGenerator()
ce = generate_counterfactual(x, target, counterfactual_data, M, generator)

animate_path(ce)

CounterfactualExplanations.converged — Method

converged(counterfactual_explanation::CounterfactualExplanation)

A convenience method to determine if the counterfactual search has converged. The search is considered to have converged only if the counterfactual is valid.

CounterfactualExplanations.generate_counterfactual — Method

generate_counterfactual(
    x::Union{AbstractArray,Int}, target::RawTargetType, data::CounterfactualData, M::Models.AbstractFittedModel, generator::AbstractGenerator;
    γ::AbstractFloat=0.75, T=1000
)

The core function that is used to run counterfactual search for a given factual x, target, counterfactual data, model and generator. Keywords can be used to specify the desired threshold for the predicted target class probability and the maximum number of iterations.

Examples

Generic generator

using CounterfactualExplanations

# Data:
using CounterfactualExplanations.Data
using Random
Random.seed!(1234)
xs, ys = Data.toy_data_linear()
X = hcat(xs...)
counterfactual_data = CounterfactualData(X,ys')

# Model
using CounterfactualExplanations.Models: LogisticModel, probs 
# Logit model:
w = [1.0 1.0] # true coefficients
b = 0
M = LogisticModel(w, [b])

# Randomly selected factual:
x = select_factual(counterfactual_data,rand(1:size(X)[2]))
y = round(probs(M, x)[1])
target = round(probs(M, x)[1])==0 ? 1 : 0 

# Counterfactual search:
generator = GenericGenerator()
counterfactual = generate_counterfactual(x, target, counterfactual_data, M, generator)

CounterfactualExplanations.get_target_index — Method

get_target_index(y_levels, target)

Utility that returns the index of target in y_levels.

CounterfactualExplanations.path — Method

path(counterfactual_explanation::CounterfactualExplanation)

A convenience method that returns the entire counterfactual path.

CounterfactualExplanations.target_probs — Function

target_probs(
    counterfactual_explanation::CounterfactualExplanation,
    x::Union{AbstractArray,Nothing}=nothing,
)

Returns the predicted probability of the target class for x. If x is nothing, the predicted probability corresponding to the counterfactual value is returned.

CounterfactualExplanations.terminated — Method

terminated(counterfactual_explanation::CounterfactualExplanation)

A convenience method to determine if the counterfactual search has terminated.

CounterfactualExplanations.total_steps — Method

total_steps(counterfactual_explanation::CounterfactualExplanation)

A convenience method that returns the total number of steps of the counterfactual search.

CounterfactualExplanations.update! — Method

update!(counterfactual_explanation::CounterfactualExplanation)

An important subroutine that updates the counterfactual explanation. It takes a snapshot of the current counterfactual search state and passes it to the generator. Based on the current state the generator generates perturbations. Various constraints are then applied to the proposed vector of feature perturbations. Finally, the counterfactual search state is updated.

CounterfactualExplanations.Evaluation.default_measures — Constant

The default evaluation measures.

CounterfactualExplanations.Evaluation.Benchmark — Type

A container for benchmarks of counterfactual explations.

CounterfactualExplanations.Evaluation.Benchmark — Method

(bmk::Benchmark)(; agg=mean)

Returns a DataFrame containing evaluation measures aggregated by num_counterfactual.

CounterfactualExplanations.Evaluation.benchmark — Method

benchmark(
    data::CounterfactualData;
    models::Dict{Symbol,Any}=model_catalogue,
    generators::Union{Nothing,Dict{<:Any,<:AbstractGenerator}}=nothing,
    measure::Union{Function,Vector{Function}}=default_measures,
    n_individuals::Int=5,
    kwrgs...
)

Runs the benchmarking exercise as follows:

Randomly choose a factual and target label.

If no pretrained models are provided, it is assumed that a dictionary of callable model objects is provided (by default using the model_catalogue).

Each of these models is then trained on the data.

For each model separately choose n_individuals randomly from the non-target (factual) class. For each generator create a benchmark as in benchmark(x::Union{AbstractArray,Base.Iterators.Zip},...).

Finally, concatenate the results.

CounterfactualExplanations.Evaluation.benchmark — Method

benchmark(
    x::Union{AbstractArray,Base.Iterators.Zip},
    target::RawTargetType,
    data::CounterfactualData;
    models::Dict{<:Any, <:AbstractFittedModel},
    generators::Dict{<:Any, <:AbstractGenerator},
    measure::Union{Function,Vector{Function}}=default_measures,
    kwrgs...
)

First generates counterfactual explanations for factual x, the target and data using each of the provided models and generators. Then generates a Benchmark for the vector of counterfactual explanations as in benchmark(counterfactual_explanations::Vector{CounterfactualExplanation}).

CounterfactualExplanations.Evaluation.benchmark — Method

benchmark(
    counterfactual_explanations::Vector{CounterfactualExplanation};
    meta_data::Union{Nothing,<:Vector{<:Dict}}=nothing,
    measure::Union{Function,Vector{Function}}=default_measures
)

Generates a Benchmark for a vector of counterfactual explanations. Optionally meta_data describing each individual counterfactual explanation can be supplied. This should be a vector of dictionaries of the same length as the vector of counterfactuals. If no meta_data is supplied, it will be automatically inferred.

CounterfactualExplanations.Evaluation.distance — Function

distance(counterfactual_explanation::CounterfactualExplanation, p::Real=2)

Computes the distance of the counterfactual to the original factual.

CounterfactualExplanations.Evaluation.evaluate — Method

evaluate(
    counterfactual_explanation::CounterfactualExplanation;
    measure::Union{Function,Vector{Function}}=default_measures,
    agg::Function=mean,
    report_each::Bool=false,
    output_format::Symbol=:Vector,
    pivot_longer::Bool=true
)

Just computes evaluation measures for the counterfactual explanation.

CounterfactualExplanations.Evaluation.evaluate — Method

evaluate(
    counterfactual_explanations::Vector{CounterfactualExplanation};
    report_meta::Bool=false,
    meta_data::Union{Nothing,<:Vector{<:Dict}}=nothing,
    kwargs...
)

Computes evaluation measures for a vector of counterfactual explanations. By default, no meta data is reported. For report_meta=true, meta data is automatically inferred, unless this overwritted by meta_data. The optional meta_data argument should be a vector of dictionaries of the same length as the vector of counterfactual explanations.

Additional kwargs... can be provided (see evaluate(counterfactual_explanation::CounterfactualExplanation for details).

CounterfactualExplanations.Evaluation.redundancy — Method

redundancy(counterfactual_explanation::CounterfactualExplanation)

Computes the feature redundancy: that is, the number of features that remain unchanged from their original, factual values.

CounterfactualExplanations.Evaluation.validity — Method

validity(counterfactual_explanation::CounterfactualExplanation; γ=0.5)

Checks of the counterfactual search has been successful with respect to the probability treshold γ. In case multiple counterfactuals were generated, the function returns the proportion of successful counterfactuals.

CounterfactualExplanations.Data.data_catalogue — Constant

A dictionary that provides an overview of the various benchmark datasets and the methods to load them.

CounterfactualExplanations.Data.load_blobs — Function

load_blobs(n=100; seed=data_seed, kwrgs...)

Loads overlapping synthtetic data.

CounterfactualExplanations.Data.load_california_housing — Function

load_california_housing(n::Union{Nothing,Int}=5000)

Loads and pre-processes California Housing data.

CounterfactualExplanations.Data.load_circles — Function

load_circles(n=100; seed=data_seed, noise=0.15, factor=0.01)

Loads synthetic circles data.

CounterfactualExplanations.Data.load_credit_default — Function

load_credit_default(n::Union{Nothing,Int}=5000)

Loads and pre-processes UCI Credit Default data.

CounterfactualExplanations.Data.load_gmsc — Function

load_gmsc(n::Union{Nothing,Int}=5000)

Loads and pre-processes Give Me Some Credit (GMSC) data.

CounterfactualExplanations.Data.load_linearly_separable — Function

load_linearly_separable(n=100; seed=data_seed)

Loads linearly separable synthtetic data.

CounterfactualExplanations.Data.load_mnist — Function

load_mnist()

Loads and prepares MNIST data.

CounterfactualExplanations.Data.load_mnist_test — Method

load_mnist_test()

Loads and prepares MNIST test data.

CounterfactualExplanations.Data.load_moons — Function

load_moons(n=100; seed=data_seed, kwrgs...)

Loads synthetic moons data.

CounterfactualExplanations.Data.load_multi_class — Function

load_multi_class(n=100; seed=data_seed)

Loads multi-class synthtetic data.

CounterfactualExplanations.Data.load_overlapping — Function

load_overlapping(n=100; seed=data_seed)

Loads overlapping synthtetic data.

CounterfactualExplanations.Data.load_synthetic_data — Function

load_synthetic_data(n=100; seed=data_seed)

Loads all synthetic datasets and wraps them in a dictionary.

CounterfactualExplanations.DataPreprocessing.CounterfactualData — Method

CounterfactualData(
    X::AbstractMatrix, y::AbstractMatrix;
    mutability::Union{Vector{Symbol},Nothing}=nothing,
    domain::Union{Any,Nothing}=nothing,
    features_categorical::Union{Vector{Int},Nothing}=nothing,
    features_continuous::Union{Vector{Int},Nothing}=nothing,
    standardize::Bool=false
)

This outer constructor method prepares features X and labels y to be used with the package. Mutability and domain constraints can be added for the features. The function also accepts arguments that specify which features are categorical and which are continues. These arguments are currently not used.

Examples

using CounterfactualExplanations.Data
x, y = toy_data_linear()
X = hcat(x...)
counterfactual_data = CounterfactualData(X,y')

CounterfactualExplanations.DataPreprocessing.CounterfactualData — Method

function CounterfactualData(
    X::Tables.MatrixTable,
    y::RawOutputArrayType;
    kwrgs...
)

Outer constructor method that accepts a Tables.MatrixTable. By default, the indices of categorical and continuous features are automatically inferred the features' scitype.

CounterfactualExplanations.DataPreprocessing.apply_domain_constraints — Method

apply_domain_constraints(counterfactual_data::CounterfactualData, x::AbstractArray)

A subroutine that is used to apply the predetermined domain constraints.

CounterfactualExplanations.DataPreprocessing.select_factual — Method

select_factual(counterfactual_data::CounterfactualData, index::Int)

A convenience method that can be used to access the the feature matrix.

CounterfactualExplanations.DataPreprocessing.transformable_features — Method

transformable_features(counterfactual_data::CounterfactualData)

Returns the indices of all continuous features that can be transformed. For constant features ZScoreTransform returns NaN.

CounterfactualExplanations.Models.model_catalogue — Constant

model_catalogue

A dictionary containing all trainable machine learning models.

CounterfactualExplanations.Models.AbstractDifferentiableModel — Type

Base type for differentiable models.

CounterfactualExplanations.Models.AbstractFittedModel — Type

Base type for fitted models.

CounterfactualExplanations.Models.FluxEnsemble — Type

FluxEnsemble <: AbstractDifferentiableJuliaModel

Constructor for deep ensembles trained in Flux.jl.

CounterfactualExplanations.Models.FluxModel — Type

FluxModel <: AbstractDifferentiableJuliaModel

Constructor for models trained in Flux.jl.

CounterfactualExplanations.Models.FluxModel — Method

FluxModel(data::CounterfactualData; kwargs...)

Constructs a multi-layer perceptron (MLP).

CounterfactualExplanations.Models.LaplaceReduxModel — Type

LaplaceReduxModel <: AbstractDifferentiableJuliaModel

Constructor for LaplaceRedux.jl model.

CounterfactualExplanations.Models.Linear — Method

Linear(data::CounterfactualData; kwargs...)

Constructs a model with one linear layer. If the output is binary, this corresponds to logistic regression, since model outputs are passed through the sigmoid function. If the output is multi-class, this corresponds to multinomial logistic regression, since model outputs are passed through the softmax function.

CounterfactualExplanations.Models.fit_model — Function

fit_model(
    counterfactual_data::CounterfactualData, model::Symbol=:MLP;
    kwrgs...
)

Fits one of the available default models to the counterfactual_data. The model argument can be used to specify the desired model. The available values correspond to the keys of the model_catalogue dictionary.

CounterfactualExplanations.Models.logits — Method

logits(M::AbstractFittedModel, X::AbstractArray)

Generic method that is compulsory for all models. It returns the raw model predictions. In classification this is sometimes referred to as logits: the non-normalized predictions that are fed into a link function to produce predicted probabilities. In regression (not currently implemented) raw outputs typically correspond to final outputs. In other words, there is typically no normalization involved.

CounterfactualExplanations.Models.model_evaluation — Method

model_evaluation(M::AbstractFittedModel, test_data::CounterfactualData)

Helper function to compute F-Score for AbstractFittedModel on a (test) data set.

CounterfactualExplanations.Models.predict_label — Method

predict_label(M::AbstractFittedModel, counterfactual_data::CounterfactualData, X::AbstractArray)

Returns the predicted output label for a given model M, data set counterfactual_data and input data X.

CounterfactualExplanations.Models.predict_label — Method

predict_label(M::AbstractFittedModel, counterfactual_data::CounterfactualData)

Returns the predicted output labels for all data points of data set counterfactual_data for a given model M.

CounterfactualExplanations.Models.predict_proba — Method

predict_proba(M::AbstractFittedModel, counterfactual_data::CounterfactualData, X::Union{Nothing,AbstractArray})

Returns the predicted output probabilities for a given model M, data set counterfactual_data and input data X.

CounterfactualExplanations.Models.probs — Method

probs(M::AbstractFittedModel, X::AbstractArray)

Generic method that is compulsory for all models. It returns the normalized model predictions, so the predicted probabilities in the case of classifiation. In regression (not currently implemented) this method is redundant.

CounterfactualExplanations.Generators.generator_catalogue — Constant

A dictionary countaining the contructors of all available counterfactual generators.

CounterfactualExplanations.Generators.AbstractGenerator — Type

AbstractGenerator

An abstract type that serves as the base type for counterfactual generators.

CounterfactualExplanations.Generators.AbstractGradientBasedGenerator — Type

AbstractGradientBasedGenerator

An abstract type that serves as the base type for gradient-based counterfactual generators.

CounterfactualExplanations.Generators.ClaPROARGenerator — Method

ClaPROARGenerator(
    ;
    loss::Symbol=:logitbinarycrossentropy,
    complexity::Function=norm,
    λ::AbstractFloat=0.1,
    opt::Flux.Optimise.AbstractOptimiser=Flux.Optimise.Descent(),
    τ::AbstractFloat=1e-5
)

An outer constructor method that instantiates a generic generator.

Examples

generator = ClaPROARGenerator()

CounterfactualExplanations.Generators.DiCEGenerator — Method

DiCEGenerator(
    ;
    loss::Symbol=:logitbinarycrossentropy,
    complexity::Function=LinearAlgebra.norm,
    λ::AbstractFloat=0.1,
    opt::Flux.Optimise.AbstractOptimiser=Flux.Optimise.Descent(),
    τ::AbstractFloat=1e-5
)

An outer constructor method that instantiates a generic generator.

Examples

generator = DiCEGenerator()

CounterfactualExplanations.Generators.GenericGenerator — Method

GenericGenerator(
    ;
    loss::Symbol=:logitbinarycrossentropy,
    complexity::Function=LinearAlgebra.norm,
    λ::AbstractFloat=0.1,
    opt::Flux.Optimise.AbstractOptimiser=Flux.Optimise.Descent(),
    τ::AbstractFloat=1e-5
)

An outer constructor method that instantiates a generic generator.

Examples

generator = GenericGenerator()

CounterfactualExplanations.Generators.GravitationalGenerator — Method

GravitationalGenerator(
    ;
    loss::Symbol=:logitbinarycrossentropy,
    complexity::Function=norm,
    λ::AbstractFloat=0.1,
    opt::Flux.Optimise.AbstractOptimiser=Flux.Optimise.Descent(),
    τ::AbstractFloat=1e-5
)

An outer constructor method that instantiates a generic generator.

Examples

generator = GravitationalGenerator()

CounterfactualExplanations.Generators.GreedyGenerator — Method

GreedyGenerator(;
    loss::Union{Nothing,Symbol}=nothing,
    complexity::Function=LinearAlgebra.norm,
    λ::AbstractFloat=0.0,
    decision_threshold=0.5,
    opt::Union{Nothing,Flux.Optimise.AbstractOptimiser}, # learning rate
    kwargs...
)

An outer constructor method that instantiates a greedy generator.

Examples

generator = GreedyGenerator()

CounterfactualExplanations.Generators.REVISEGenerator — Method

REVISEGenerator(
    ;
    loss::Symbol=:logitbinarycrossentropy,
    complexity::Function=LinearAlgebra.norm,
    λ::AbstractFloat=0.1,
    opt::Flux.Optimise.AbstractOptimiser=Flux.Optimise.Descent(),
    τ::AbstractFloat=1e-5
)

An outer constructor method that instantiates a REVISE generator.

Examples

generator = REVISEGenerator()

CounterfactualExplanations.Generators.conditions_satisified — Method

conditions_satisified(generator::AbstractGradientBasedGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to check if the all conditions for convergence of the counterfactual search have been satisified for gradient-based generators. By default, gradient-based search is considered to have converged as soon as the proposed feature changes for all features are smaller than one percent of its standard deviation.

CounterfactualExplanations.Generators.generate_perturbations — Method

generate_perturbations(generator::AbstractGradientBasedGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to generate feature perturbations for gradient-based generators through simple gradient descent.

CounterfactualExplanations.Generators.generate_perturbations — Method

generate_perturbations(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to generate perturbations for a greedy generator. Only the most salient feature is perturbed.

CounterfactualExplanations.Generators.mutability_constraints — Method

mutability_constraints(generator::AbstractGradientBasedGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to return mutability constraints that are dependent on the current counterfactual search state. For generic gradient-based generators, no state-dependent constraints are added.

CounterfactualExplanations.Generators.mutability_constraints — Method

mutability_constraints(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to return search state dependent mutability constraints for a greedy generator. Features that have been perturbed n times already can no longer be perturbed.

Flux.Losses.binary_focal_loss — Method

binary_focal_loss(ŷ, y; agg=mean, γ=2, ϵ=eps(ŷ))

Return the binaryfocalloss The input, 'ŷ', is expected to be normalized (i.e. softmax output).

For γ == 0, the loss is mathematically equivalent to Losses.binarycrossentropy.

See also: Losses.focal_loss for multi-class setting

Example

julia> y = [0  1  0
            1  0  1]
2×3 Matrix{Int64}:
 0  1  0
 1  0  1

julia> ŷ = [0.268941  0.5  0.268941
            0.731059  0.5  0.731059]
2×3 Matrix{Float64}:
 0.268941  0.5  0.268941
 0.731059  0.5  0.731059

julia> Flux.binary_focal_loss(ŷ, y) ≈ 0.0728675615927385
true

Flux.Losses.binarycrossentropy — Method

binarycrossentropy(ŷ, y; agg = mean, ϵ = eps(ŷ))

Return the binary cross-entropy loss, computed as

agg(@.(-y * log(ŷ + ϵ) - (1 - y) * log(1 - ŷ + ϵ)))

Where typically, the prediction ŷ is given by the output of a sigmoid activation. The ϵ term is included to avoid infinity. Using logitbinarycrossentropy is recomended over binarycrossentropy for numerical stability.

Use label_smoothing to smooth the y value as preprocessing before computing the loss.

See also: crossentropy, logitcrossentropy.

Examples

julia> y_bin = Bool[1,0,1]
3-element Vector{Bool}:
 1
 0
 1

julia> y_prob = softmax(reshape(vcat(1:3, 3:5), 2, 3) .* 1f0)
2×3 Matrix{Float32}:
 0.268941  0.5  0.268941
 0.731059  0.5  0.731059

julia> Flux.binarycrossentropy(y_prob[2,:], y_bin)
0.43989f0

julia> all(p -> 0 < p < 1, y_prob[2,:])  # else DomainError
true

julia> y_hot = Flux.onehotbatch(y_bin, 0:1)
2×3 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
 ⋅  1  ⋅
 1  ⋅  1

julia> Flux.crossentropy(y_prob, y_hot)
0.43989f0

Flux.Losses.crossentropy — Method

crossentropy(ŷ, y; dims = 1, ϵ = eps(ŷ), agg = mean)

Return the cross entropy between the given probability distributions; calculated as

agg(-sum(y .* log.(ŷ .+ ϵ); dims))

Cross entropy is typically used as a loss in multi-class classification, in which case the labels y are given in a one-hot format. dims specifies the dimension (or the dimensions) containing the class probabilities. The prediction ŷ is supposed to sum to one across dims, as would be the case with the output of a softmax operation.

For numerical stability, it is recommended to use logitcrossentropy rather than softmax followed by crossentropy .

Use label_smoothing to smooth the true labels as preprocessing before computing the loss.

See also: logitcrossentropy, binarycrossentropy, logitbinarycrossentropy.

Example

julia> y_label = Flux.onehotbatch([0, 1, 2, 1, 0], 0:2)
3×5 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
 1  ⋅  ⋅  ⋅  1
 ⋅  1  ⋅  1  ⋅
 ⋅  ⋅  1  ⋅  ⋅

julia> y_model = softmax(reshape(-7:7, 3, 5) .* 1f0)
3×5 Matrix{Float32}:
 0.0900306  0.0900306  0.0900306  0.0900306  0.0900306
 0.244728   0.244728   0.244728   0.244728   0.244728
 0.665241   0.665241   0.665241   0.665241   0.665241

julia> sum(y_model; dims=1)
1×5 Matrix{Float32}:
 1.0  1.0  1.0  1.0  1.0

julia> Flux.crossentropy(y_model, y_label)
1.6076053f0

julia> 5 * ans ≈ Flux.crossentropy(y_model, y_label; agg=sum)
true

julia> y_smooth = Flux.label_smoothing(y_label, 0.15f0)
3×5 Matrix{Float32}:
 0.9   0.05  0.05  0.05  0.9
 0.05  0.9   0.05  0.9   0.05
 0.05  0.05  0.9   0.05  0.05

julia> Flux.crossentropy(y_model, y_smooth)
1.5776052f0

Flux.Losses.dice_coeff_loss — Method

dice_coeff_loss(ŷ, y; smooth = 1)

Return a loss based on the dice coefficient. Used in the V-Net image segmentation architecture. The dice coefficient is similar to the F1_score. Loss calculated as:

1 - 2*sum(|ŷ .* y| + smooth) / (sum(ŷ.^2) + sum(y.^2) + smooth)

Example

julia> y_pred = [1.1, 2.1, 3.1];

julia> Flux.dice_coeff_loss(y_pred, 1:3)
0.000992391663909964

julia> 1 - Flux.dice_coeff_loss(y_pred, 1:3)  # ~ F1 score for image segmentation
0.99900760833609

Flux.Losses.focal_loss — Method

focal_loss(ŷ, y; dims=1, agg=mean, γ=2, ϵ=eps(ŷ))

Return the focal_loss which can be used in classification tasks with highly imbalanced classes. It down-weights well-classified examples and focuses on hard examples. The input, 'ŷ', is expected to be normalized (i.e. softmax output).

The modulating factor, γ, controls the down-weighting strength. For γ == 0, the loss is mathematically equivalent to Losses.crossentropy.

Example

julia> y = [1  0  0  0  1
            0  1  0  1  0
            0  0  1  0  0]
3×5 Matrix{Int64}:
 1  0  0  0  1
 0  1  0  1  0
 0  0  1  0  0

julia> ŷ = softmax(reshape(-7:7, 3, 5) .* 1f0)
3×5 Matrix{Float32}:
 0.0900306  0.0900306  0.0900306  0.0900306  0.0900306
 0.244728   0.244728   0.244728   0.244728   0.244728
 0.665241   0.665241   0.665241   0.665241   0.665241

julia> Flux.focal_loss(ŷ, y) ≈ 1.1277571935622628
true

See also: Losses.binary_focal_loss for binary (not one-hot) labels

Flux.Losses.hinge_loss — Method

hinge_loss(ŷ, y; agg = mean)

Return the hinge_loss given the prediction ŷ and true labels y (containing 1 or -1); calculated as

sum(max.(0, 1 .- ŷ .* y)) / size(y, 2)

Usually used with classifiers like Support Vector Machines. See also: squared_hinge_loss

Example

julia> y_true = [1, -1, 1, 1];

julia> y_pred = [0.1, 0.3, 1, 1.5];

julia> Flux.hinge_loss(y_pred, y_true)
0.55

julia> Flux.hinge_loss(y_pred[1], y_true[1]) != 0  # same sign but |ŷ| < 1
true

julia> Flux.hinge_loss(y_pred[end], y_true[end]) == 0  # same sign but |ŷ| >= 1
true

julia> Flux.hinge_loss(y_pred[2], y_true[2]) != 0 # opposite signs
true

Flux.Losses.huber_loss — Method

huber_loss(ŷ, y; δ = 1, agg = mean)

Return the mean of the Huber loss given the prediction ŷ and true values y.

             | 0.5 * |ŷ - y|^2,            for |ŷ - y| <= δ
Huber loss = |
             |  δ * (|ŷ - y| - 0.5 * δ), otherwise

Example

julia> ŷ = [1.1, 2.1, 3.1];

julia> Flux.huber_loss(ŷ, 1:3)  # default δ = 1 > |ŷ - y|
0.005000000000000009

julia> Flux.huber_loss(ŷ, 1:3, δ=0.05)  # changes behaviour as |ŷ - y| > δ
0.003750000000000005

Flux.Losses.kldivergence — Method

kldivergence(ŷ, y; agg = mean, ϵ = eps(ŷ))

Return the Kullback-Leibler divergence between the given probability distributions.

The KL divergence is a measure of how much one probability distribution is different from the other. It is always non-negative, and zero only when both the distributions are equal.

Example

julia> p1 = [1 0; 0 1]
2×2 Matrix{Int64}:
 1  0
 0  1

julia> p2 = fill(0.5, 2, 2)
2×2 Matrix{Float64}:
 0.5  0.5
 0.5  0.5

julia> Flux.kldivergence(p2, p1) ≈ log(2)
true

julia> Flux.kldivergence(p2, p1; agg = sum) ≈ 2log(2)
true

julia> Flux.kldivergence(p2, p2; ϵ = 0)  # about -2e-16 with the regulator
0.0

julia> Flux.kldivergence(p1, p2; ϵ = 0)  # about 17.3 with the regulator
Inf

Flux.Losses.label_smoothing — Method

label_smoothing(y::Union{Number, AbstractArray}, α; dims::Int=1)

Returns smoothed labels, meaning the confidence on label values are relaxed.

When y is given as one-hot vector or batch of one-hot, its calculated as

y .* (1 - α) .+ α / size(y, dims)

when y is given as a number or batch of numbers for binary classification, its calculated as

y .* (1 - α) .+ α / 2

in which case the labels are squeezed towards 0.5.

α is a number in interval (0, 1) called the smoothing factor. Higher the value of α larger the smoothing of y.

dims denotes the one-hot dimension, unless dims=0 which denotes the application of label smoothing to binary distributions encoded in a single number.

Example

julia> y = Flux.onehotbatch([1, 1, 1, 0, 1, 0], 0:1)
2×6 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
 ⋅  ⋅  ⋅  1  ⋅  1
 1  1  1  ⋅  1  ⋅

julia> y_smoothed = Flux.label_smoothing(y, 0.2f0)
2×6 Matrix{Float32}:
 0.1  0.1  0.1  0.9  0.1  0.9
 0.9  0.9  0.9  0.1  0.9  0.1

julia> y_sim = softmax(y .* log(2f0))
2×6 Matrix{Float32}:
 0.333333  0.333333  0.333333  0.666667  0.333333  0.666667
 0.666667  0.666667  0.666667  0.333333  0.666667  0.333333

julia> y_dis = vcat(y_sim[2,:]', y_sim[1,:]')
2×6 Matrix{Float32}:
 0.666667  0.666667  0.666667  0.333333  0.666667  0.333333
 0.333333  0.333333  0.333333  0.666667  0.333333  0.666667

julia> Flux.crossentropy(y_sim, y) < Flux.crossentropy(y_sim, y_smoothed)
true

julia> Flux.crossentropy(y_dis, y) > Flux.crossentropy(y_dis, y_smoothed)
true

Flux.Losses.logitbinarycrossentropy — Method

logitbinarycrossentropy(ŷ, y; agg = mean)

Mathematically equivalent to binarycrossentropy(σ(ŷ), y) but is more numerically stable.

See also: crossentropy, logitcrossentropy.

Examples

julia> y_bin = Bool[1,0,1];

julia> y_model = Float32[2, -1, pi]
3-element Vector{Float32}:
  2.0
 -1.0
  3.1415927

julia> Flux.logitbinarycrossentropy(y_model, y_bin)
0.160832f0

julia> Flux.binarycrossentropy(sigmoid.(y_model), y_bin)
0.16083185f0

Flux.Losses.logitcrossentropy — Method

logitcrossentropy(ŷ, y; dims = 1, agg = mean)

Return the cross entropy calculated by

agg(-sum(y .* logsoftmax(ŷ; dims); dims))

This is mathematically equivalent to crossentropy(softmax(ŷ), y), but is more numerically stable than using functions crossentropy and softmax separately.

See also: binarycrossentropy, logitbinarycrossentropy, label_smoothing.

Example

julia> y_label = Flux.onehotbatch(collect("abcabaa"), 'a':'c')
3×7 OneHotMatrix(::Vector{UInt32}) with eltype Bool:
 1  ⋅  ⋅  1  ⋅  1  1
 ⋅  1  ⋅  ⋅  1  ⋅  ⋅
 ⋅  ⋅  1  ⋅  ⋅  ⋅  ⋅

julia> y_model = reshape(vcat(-9:0, 0:9, 7.5f0), 3, 7)
3×7 Matrix{Float32}:
 -9.0  -6.0  -3.0  0.0  2.0  5.0  8.0
 -8.0  -5.0  -2.0  0.0  3.0  6.0  9.0
 -7.0  -4.0  -1.0  1.0  4.0  7.0  7.5

julia> Flux.logitcrossentropy(y_model, y_label)
1.5791205f0

julia> Flux.crossentropy(softmax(y_model), y_label)
1.5791197f0

Flux.Losses.mae — Method

mae(ŷ, y; agg = mean)

Return the loss corresponding to mean absolute error:

agg(abs.(ŷ .- y))

Example

julia> y_model = [1.1, 1.9, 3.1];

julia> Flux.mae(y_model, 1:3)
0.10000000000000009

Flux.Losses.mse — Method

mse(ŷ, y; agg = mean)

Return the loss corresponding to mean square error:

agg((ŷ .- y) .^ 2)

See also: mae, msle, crossentropy.

Example

julia> y_model = [1.1, 1.9, 3.1];

julia> y_true = 1:3;

julia> Flux.mse(y_model, y_true)
0.010000000000000018

Flux.Losses.msle — Method

msle(ŷ, y; agg = mean, ϵ = eps(ŷ))

The loss corresponding to mean squared logarithmic errors, calculated as

agg((log.(ŷ .+ ϵ) .- log.(y .+ ϵ)) .^ 2)

The ϵ term provides numerical stability. Penalizes an under-estimation more than an over-estimatation.

Example

julia> Flux.msle(Float32[1.1, 2.2, 3.3], 1:3)
0.009084041f0

julia> Flux.msle(Float32[0.9, 1.8, 2.7], 1:3)
0.011100831f0

Flux.Losses.poisson_loss — Method

poisson_loss(ŷ, y; agg = mean)

Return how much the predicted distribution ŷ diverges from the expected Poisson distribution y; calculated as -

sum(ŷ .- y .* log.(ŷ)) / size(y, 2)

More information..

Example

julia> y_model = [1, 3, 3];  # data should only take integral values

julia> Flux.poisson_loss(y_model, 1:3)
0.5023128522198171

Flux.Losses.siamese_contrastive_loss — Method

siamese_contrastive_loss(ŷ, y; margin = 1, agg = mean)

Return the contrastive loss which can be useful for training Siamese Networks. It is given by

agg(@. (1 - y) * ŷ^2 + y * max(0, margin - ŷ)^2)

Specify margin to set the baseline for distance at which pairs are dissimilar.

Example

julia> ŷ = [0.5, 1.5, 2.5];

julia> Flux.siamese_contrastive_loss(ŷ, 1:3)
-4.833333333333333

julia> Flux.siamese_contrastive_loss(ŷ, 1:3, margin = 2)
-4.0

Flux.Losses.squared_hinge_loss — Method

squared_hinge_loss(ŷ, y)

Return the squared hinge_loss loss given the prediction ŷ and true labels y (containing 1 or -1); calculated as

sum((max.(0, 1 .- ŷ .* y)).^2) / size(y, 2)

Usually used with classifiers like Support Vector Machines. See also: hinge_loss

Example

julia> y_true = [1, -1, 1, 1];

julia> y_pred = [0.1, 0.3, 1, 1.5];

julia> Flux.squared_hinge_loss(y_pred, y_true)
0.625

julia> Flux.squared_hinge_loss(y_pred[1], y_true[1]) != 0
true

julia> Flux.squared_hinge_loss(y_pred[end], y_true[end]) == 0
true

julia> Flux.squared_hinge_loss(y_pred[2], y_true[2]) != 0
true

Flux.Losses.tversky_loss — Method

tversky_loss(ŷ, y; β = 0.7)

Return the Tversky loss. Used with imbalanced data to give more weight to false negatives. Larger β weigh recall more than precision (by placing more emphasis on false negatives). Calculated as:

1 - sum(|y .* ŷ| + 1) / (sum(y .* ŷ + (1 - β)*(1 .- y) .* ŷ + β*y .* (1 .- ŷ)) + 1)

Internal functions

CounterfactualExplanations.FluxModelParams — Type

FluxModelParams

Default MLP training parameters.

CounterfactualExplanations.PlotIngredients — Type

A container used for plotting.

CounterfactualExplanations.adjust_shape! — Method

adjust_shape!(counterfactual_explanation::CounterfactualExplanation)

A convenience method that adjusts the dimensions of the counterfactual state and related fields.

CounterfactualExplanations.adjust_shape — Method

adjust_shape(
    counterfactual_explanation::CounterfactualExplanation, 
    x::AbstractArray
)

A convenience method that adjusts the dimensions of x.

CounterfactualExplanations.apply_domain_constraints! — Method

apply_domain_constraints!(counterfactual_explanation::CounterfactualExplanation)

Wrapper function that applies underlying domain constraints.

CounterfactualExplanations.apply_mutability — Method

apply_mutability(
    counterfactual_explanation::CounterfactualExplanation,
    Δs′::AbstractArray,
)

A subroutine that applies mutability constraints to the proposed vector of feature perturbations.

CounterfactualExplanations.counterfactual — Method

counterfactual(counterfactual_explanation::CounterfactualExplanation)

A convenience method that returns the counterfactual.

CounterfactualExplanations.counterfactual_label — Method

counterfactual_label(counterfactual_explanation::CounterfactualExplanation)

A convenience method that returns the predicted label of the counterfactual.

CounterfactualExplanations.counterfactual_label_path — Method

counterfactual_label_path(counterfactual_explanation::CounterfactualExplanation)

Returns the counterfactual labels for each step of the search.

CounterfactualExplanations.counterfactual_probability — Method

counterfactual_probability(counterfactual_explanation::CounterfactualExplanation)

A convenience method that computes the class probabilities of the counterfactual.

CounterfactualExplanations.counterfactual_probability_path — Method

counterfactual_probability_path(counterfactual_explanation::CounterfactualExplanation)

Returns the counterfactual probabilities for each step of the search.

CounterfactualExplanations.decode_state — Function

function decode_state(
    counterfactual_explanation::CounterfactualExplanation,
    x::Union{AbstractArray,Nothing}=nothing,
)

Applies all the applicable decoding functions:

If applicable, map the state variable back from the latent space to the feature space.

If and where applicable, inverse-transform features.

Reconstruct all categorical encodings.

Finally, the decoded counterfactual is returned.

CounterfactualExplanations.embed_path — Method

embed_path(counterfactual_explanation::CounterfactualExplanation)

Helper function that embeds path into two dimensions for plotting.

CounterfactualExplanations.encode_state — Function

function encode_state(
    counterfactual_explanation::CounterfactualExplanation, 
    x::Union{AbstractArray,Nothing} = nothing,
)

Applies all required encodings to x:

If applicable, it maps x to the latent space learned by the generative model.

If and where applicable, it rescales features.

Finally, it returns the encoded state variable.

CounterfactualExplanations.factual — Method

factual(counterfactual_explanation::CounterfactualExplanation)

A convenience method to retrieve the factual x.

CounterfactualExplanations.factual_label — Method

factual_label(counterfactual_explanation::CounterfactualExplanation)

A convenience method to get the predicted label associated with the factual.

CounterfactualExplanations.factual_probability — Method

factual_probability(counterfactual_explanation::CounterfactualExplanation)

A convenience method to compute the class probabilities of the factual.

CounterfactualExplanations.get_meta — Method

get_meta(counterfactual_explanation::CounterfactualExplanation)

Returns meta data for a counterfactual explanation.

CounterfactualExplanations.guess_likelihood — Method

guess_likelihood(y::RawOutputArrayType)

Guess the likelihood based on the scientific type of the output array. Returns a symbol indicating the guessed likelihood and the scientific type of the output array.

CounterfactualExplanations.guess_loss — Method

guess_loss(counterfactual_explanation::CounterfactualExplanation)

Guesses the loss function to be used for the counterfactual search in case likelihood field is specified for the AbstractFittedModel instance and no loss function was explicitly declared for AbstractGenerator instance.

CounterfactualExplanations.initialize_state — Method

initialize_state(counterfactual_explanation::CounterfactualExplanation)

Initializes the starting point for the factual(s):

If counterfactual_explanation.initialization is set to :identity or counterfactuals are searched in a latent space, then nothing is done.

If counterfactual_explanation.initialization is set to :add_perturbation, then a random perturbation is added to the factual following following Slack (2021): https://arxiv.org/abs/2106.02666. The authors show that this improves adversarial robustness.

CounterfactualExplanations.map_from_latent — Function

map_from_latent(
    counterfactual_explanation::CounterfactualExplanation,
    x::Union{AbstractArray,Nothing}=nothing,
)

Maps the state variable back from the latent space to the feature space.

CounterfactualExplanations.map_to_latent — Function

function maptolatent( counterfactual_explanation::CounterfactualExplanation, x::Union{AbstractArray,Nothing}=nothing, )

Maps x from the feature space $\mathcal{X}$ to the latent space learned by the generative model.

CounterfactualExplanations.output_dim — Method

output_dim(counterfactual_explanation::CounterfactualExplanation)

A convenience method that returns the output dimension of the predictive model.

CounterfactualExplanations.plot_state — Method

plot_state(
    counterfactual_explanation::CounterfactualExplanation,
    t::Int,
    final_sate::Bool;
    kwargs...
)

Helper function that plots a single step of the counterfactual path.

CounterfactualExplanations.reconstruct_cat_encoding — Function

reconstruct_cat_encoding(
    counterfactual_explanation::CounterfactualExplanation,
    x::Union{AbstractArray,Nothing}=nothing,
)

Reconstructs all categorical encodings. See DataPreprocessing.reconstruct_cat_encoding for details.

CounterfactualExplanations.reset! — Method

reset!(flux_training_params::FluxModelParams)

Restores the default parameter values.

CounterfactualExplanations.set_up_plots — Method

set_up_plots(
    counterfactual_explanation::CounterfactualExplanation;
    alpha,
    plot_proba,
    kwargs...
)

A helper method that prepares data for plotting.

CounterfactualExplanations.steps_exhausted — Method

steps_exhausted(counterfactual_explanation::CounterfactualExplanation)

A convenience method that checks if the number of maximum iterations has been exhausted.

CounterfactualExplanations.target_probs_path — Method

target_probs_path(counterfactual_explanation::CounterfactualExplanation)

Returns the target probabilities for each step of the search.

CounterfactualExplanations.threshold_reached — Method

threshold_reached(counterfactual_explanation::CounterfactualExplanation, x::AbstractArray)

A convenience method that determines if the predefined threshold for the target class probability has been reached for a specific sample x.

CounterfactualExplanations.threshold_reached — Method

threshold_reached(counterfactual_explanation::CounterfactualExplanation)

A convenience method that determines if the predefined threshold for the target class probability has been reached.

CounterfactualExplanations.wants_latent_space — Method

wants_latent_space(
    counterfactual_explanation::CounterfactualExplanation, 
    x::Union{AbstractArray,Nothing} = nothing,
)

A convenience function that checks if latent space search is applicable.

RecipesBase.plot — Method

Plots.plot(
    counterfactual_explanation::CounterfactualExplanation;
    alpha_ = 0.5,
    plot_up_to::Union{Nothing,Int} = nothing,
    plot_proba::Bool = false,
    kwargs...,
)

Calling plot on an instance of type CounterfactualExplanation returns a plot that visualises the entire counterfactual path. For multi-dimensional input data, the data is first compressed into two dimensions. The decision boundary is then approximated using using a Nearest Neighbour classifier. This is still somewhat experimental at the moment.

Examples

# Search:
generator = GenericGenerator()
ce = generate_counterfactual(x, target, counterfactual_data, M, generator)

plot(ce)

CounterfactualExplanations.Evaluation.distance_measures — Constant

All distance measures.

Base.vcat — Method

Base.vcat(bmk1::Benchmark, bmk2::Benchmark)

Vertically concatenates two Benchmark objects.

CounterfactualExplanations.Evaluation.distance_l0 — Method

distance_l0(counterfactual_explanaation::CounterfactualExplanation)

Computes the L0 distance of the counterfactual to the original factual.

CounterfactualExplanations.Evaluation.distance_l1 — Method

distance_l1(counterfactual_explanation::CounterfactualExplanation)

Computes the L1 distance of the counterfactual to the original factual.

CounterfactualExplanations.Evaluation.distance_l2 — Method

distance_l2(counterfactual_explanation::CounterfactualExplanation)

Computes the L2 (Euclidean) distance of the counterfactual to the original factual.

CounterfactualExplanations.Evaluation.distance_linf — Method

distance_linf(counterfactual_explanation::CounterfactualExplanation)

Computes the L-inf distance of the counterfactual to the original factual.

CounterfactualExplanations.Evaluation.validity_strict — Method

validity_strict(counterfactual_explanation::CounterfactualExplanation)

Checks if the counterfactual search has been strictly valid in the sense that it has converged with respect to the pre-specified target probability γ.

CounterfactualExplanations.DataPreprocessing._subset — Method

_subset(data::CounterfactualData, idx::Vector{Int})

Creates a subset of the data.

CounterfactualExplanations.DataPreprocessing.get_generative_model — Method

get_generative_model(counterfactual_data::CounterfactualData)

Returns the underlying generative model. If there is no existing model available, the default generative model (VAE) is used. Otherwise it is expected that existing generative model has been pre-trained or else a warning is triggered.

CounterfactualExplanations.DataPreprocessing.has_pretrained_generative_model — Method

has_pretrained_generative_model(counterfactual_data::CounterfactualData)

Checks if generative model is present and trained.

CounterfactualExplanations.DataPreprocessing.input_dim — Method

input_dim(counterfactual_data::CounterfactualData)

Helper function that returns the input dimension (number of features) of the data.

CounterfactualExplanations.DataPreprocessing.mutability_constraints — Method

mutability_constraints(counterfactual_data::CounterfactualData)

A convience function that returns the mutability constraints. If none were specified, it is assumed that all features are mutable in :both directions.

CounterfactualExplanations.DataPreprocessing.reconstruct_cat_encoding — Method

reconstruct_cat_encoding(counterfactual_data::CounterfactualData, x::Vector)

Reconstruct the categorical encoding for a single instance.

CounterfactualExplanations.DataPreprocessing.train_test_split — Method

train_test_split(data::CounterfactualData;test_size=0.2)

Splits data into train and test split.

CounterfactualExplanations.DataPreprocessing.undersample — Method

undersample(data::CounterfactualData, n::Int)

Helper function to randomly undersample data::CounterfactualData.

CounterfactualExplanations.DataPreprocessing.unpack_data — Method

unpack_data(data::CounterfactualData)

Helper function that unpacks data.

CounterfactualExplanations.Models.AbstractDifferentiableJuliaModel — Type

Base type for differentiable models written in pure Julia.

CounterfactualExplanations.Models.FluxEnsembleParams — Type

FluxModelParams

Default Deep Ensemble training parameters.

CounterfactualExplanations.Models.build_ensemble — Method

build_ensemble(K::Int;kw=(input_dim=2,n_hidden=32,output_dim=1))

Helper function that builds an ensemble of K models.

CounterfactualExplanations.Models.build_mlp — Method

build_mlp()

Helper function to build simple MLP.

Examples

nn = build_mlp()

CounterfactualExplanations.Models.data_loader — Method

data_loader(data::CounterfactualData)

Prepares counterfactual data for training in Flux.

CounterfactualExplanations.Models.train — Method

train(M::FluxEnsemble, data::CounterfactualData; kwargs...)

Wrapper function to retrain.

CounterfactualExplanations.Models.train — Method

train(M::FluxModel, data::CounterfactualData; kwargs...)

Wrapper function to retrain FluxModel.

CounterfactualExplanations.Generators.AbstractLatentSpaceGenerator — Type

AbstractLatentSpaceGenerator

An abstract type that serves as the base type for gradient-based counterfactual generators that search in a latent space.

CounterfactualExplanations.Generators._replace_nans — Function

_replace_nans(Δs′::AbstractArray, old_new::Pair=(NaN => 0))

Helper function to deal with exploding gradients. This is only a temporary fix and will be improved.

CounterfactualExplanations.Generators.gradient_penalty — Method

gradient_penalty(
    generator::ClaPROARGenerator,
    counterfactual_explanation::AbstractCounterfactualExplanation,
)

Additional penalty for ClaPROARGenerator.

CounterfactualExplanations.Generators.h — Method

h(generator::AbstractGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to apply the generator complexity penalty to the current counterfactual state for any generator.

CounterfactualExplanations.Generators.h — Method

h(generator::AbstractGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to apply the generator complexity penalty to the current counterfactual state for any generator.

CounterfactualExplanations.Generators.h — Method

h(generator::AbstractGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to apply the generator complexity penalty to the current counterfactual state for any generator.

CounterfactualExplanations.Generators.h — Method

h(generator::AbstractGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to apply the generator complexity penalty to the current counterfactual state for any generator.

CounterfactualExplanations.Generators.propose_state — Method

propose_state(generator::AbstractGradientBasedGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

Proposes new state based on backpropagation.

CounterfactualExplanations.Generators.ℓ — Method

ℓ(generator::AbstractGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to apply the generator loss function to the current counterfactual state for any generator.

CounterfactualExplanations.Generators.∂h — Method

∂h(generator::AbstractGradientBasedGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to compute the gradient of the complexity penalty at the current counterfactual state for gradient-based generators. It assumes that Zygote.jl has gradient access.

CounterfactualExplanations.Generators.∂ℓ — Method

∂ℓ(generator::AbstractGradientBasedGenerator, M::Union{Models.LogisticModel, Models.BayesianLogisticModel}, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to compute the gradient of the loss function at the current counterfactual state for gradient-based generators. It assumes that Zygote.jl has gradient access.

CounterfactualExplanations.Generators.∇ — Method

∇(generator::AbstractGradientBasedGenerator, M::Models.AbstractDifferentiableModel, counterfactual_explanation::AbstractCounterfactualExplanation)

The default method to compute the gradient of the counterfactual search objective for gradient-based generators. It simply computes the weighted sum over partial derivates. It assumes that Zygote.jl has gradient access.

CounterfactualExplanations.Generators.∇ — Method

∇(generator::GreedyGenerator, counterfactual_explanation::AbstractCounterfactualExplanation)

he default method to compute the gradient of the counterfactual search objective for a greedy generator. Since no complexity penalty is needed, this gradients just correponds to the partial derivative with respect to the loss function.

Flux.Losses.xlogx — Method

xlogx(x)

Return x * log(x) for x ≥ 0, handling x == 0 by taking the limit from above, to get zero.

Flux.Losses.xlogy — Method

xlogy(x, y)

Return x * log(y) for y > 0, and zero when x == 0.